requirements.txt
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-filter
django-ordered-model
drf-nested-routers